refactor(utils.projen-builder): extract `project-builder` module
